+++ modules/router/src/forwarding.ts
@@ -37,7 +37,12 @@ export async function forwardTransferCreation(
   store: IRouterStore,
   logger: BaseLogger,
   chainReader: IVectorChainReader,
-): Promise<Result<any, ForwardTransferCreationError>> {
+): Promise<
+  Result<
+    (NodeResponses.ConditionalTransfer & { recipientAmount: string; recipientAssetId: string }) | undefined,
+    ForwardTransferCreationError
+  >
+> {
   const method = "forwardTransferCreation";
   const methodId = getRandomBytes32();
   logger.info(
@@ -211,7 +216,13 @@ export async function forwardTransferCreation(
       },
       "Detected inflight swap",
     );
-    const swapRes = getSwappedAmount(senderAmount, senderAssetId, senderChainId, recipientAssetId, recipientChainId);
+    const swapRes = await getSwappedAmount(
+      senderAmount,
+      senderAssetId,
+      senderChainId,
+      recipientAssetId,
+      recipientChainId,
+    );
     if (swapRes.isError) {
       return cancelSenderTransferAndReturnError(
         routingId,
@@ -303,7 +314,7 @@ export async function forwardTransferCreation(
     // transfer was either queued or executed
     const value = transferRes.getValue();
     return !!value
-      ? Result.ok(value)
+      ? Result.ok({ ...value, recipientAmount, recipientAssetId })
       : Result.fail(
           new ForwardTransferCreationError(
             ForwardTransferCreationError.reasons.ReceiverOffline,
@@ -463,7 +474,10 @@ export async function forwardTransferResolution(
   }
 
   logger.info({ method, methodId }, "Method complete");
-  return Result.ok({ ...resolution.getValue(), assetId: incomingTransfer.assetId });
+  return Result.ok({
+    ...resolution.getValue(),
+    assetId: incomingTransfer.assetId,
+  });
 }
 
 export async function handleIsAlive(