+++ modules/router/src/index.ts
@@ -16,20 +16,20 @@ import {
   WithdrawalCreatedPayload,
   WithdrawalReconciledPayload,
   WithdrawalResolvedPayload,
-  HydratedProviders,
-  ERC20Abi,
+  TransactionSubmittedPayload,
+  TransactionMinedPayload,
+  TransactionFailedPayload,
   SetupPayload,
 } from "@connext/vector-types";
-import { collectDefaultMetrics, Gauge, Registry } from "prom-client";
+import { collectDefaultMetrics, register } from "prom-client";
 import { Wallet } from "ethers";
 
 import { config } from "./config";
 import { IRouter, Router } from "./router";
 import { PrismaStore } from "./services/store";
 import { NatsRouterMessagingService } from "./services/messaging";
-import { AddressZero } from "@ethersproject/constants";
-import { formatEther, formatUnits } from "@ethersproject/units";
-import { Contract } from "@ethersproject/contracts";
+import { autoRebalanceTask, startAutoRebalanceTask } from "./autoRebalance";
+import { wallet } from "./metrics";
 
 const routerPort = 8000;
 const routerBase = `http://router:${routerPort}`;
@@ -43,6 +43,9 @@ const restoreStatePath = "/restore-state";
 const withdrawalCreatedPath = "/withdrawal-created";
 const withdrawReconciledPath = "/withdrawal-reconciled";
 const withdrawResolvedPath = "/withdrawal-resolved";
+const transactionSubmittedPath = "/transaction-submitted";
+const transactionMinedPath = "/transaction-mined";
+const transactionFailedPath = "/transaction-failed";
 const evts: EventCallbackConfig = {
   [EngineEvents.IS_ALIVE]: {
     evt: Evt.create<IsAlivePayload>(),
@@ -84,6 +87,18 @@ const evts: EventCallbackConfig = {
     evt: Evt.create<WithdrawalResolvedPayload>(),
     url: `${routerBase}${withdrawResolvedPath}`,
   },
+  [EngineEvents.TRANSACTION_SUBMITTED]: {
+    evt: Evt.create<TransactionSubmittedPayload & { publicIdentifier: string }>(),
+    url: `${routerBase}${transactionSubmittedPath}`,
+  },
+  [EngineEvents.TRANSACTION_MINED]: {
+    evt: Evt.create<TransactionMinedPayload & { publicIdentifier: string }>(),
+    url: `${routerBase}${transactionMinedPath}`,
+  },
+  [EngineEvents.TRANSACTION_FAILED]: {
+    evt: Evt.create<TransactionFailedPayload & { publicIdentifier: string }>(),
+    url: `${routerBase}${transactionFailedPath}`,
+  },
 };
 
 const signer = new ChannelSigner(Wallet.fromMnemonic(config.mnemonic).privateKey);
@@ -92,82 +107,25 @@ const logger = pino({ name: signer.publicIdentifier });
 logger.info({ config }, "Loaded config from environment");
 const server = fastify({ logger, pluginTimeout: 300_000, disableRequestLogging: config.logLevel !== "debug" });
 
-const register = new Registry();
-collectDefaultMetrics({ register, prefix: "router_" });
+collectDefaultMetrics({ prefix: "router_" });
 
 let router: IRouter;
 const store = new PrismaStore();
-
-const hydrated: HydratedProviders = hydrateProviders(config.chainProviders);
-
-// create gauge to store balance for each rebalanced asset and each native asset for the signer address
-// TODO: maybe want to look into an API rather than blowing up our eth providers? although its not that many calls
-
-// get all non-zero addresses
-const rebalancedTokens: {
-  [chainId: string]: {
-    [assetId: string]: {
-      contract: Contract;
-      decimals?: number;
-    };
-  };
-} = {};
-Object.entries(hydrated).forEach(async ([chainId, provider]) => {
-  rebalancedTokens[chainId] = {};
-  const assets = config.rebalanceProfiles
-    .filter((prof) => prof.chainId.toString() === chainId && prof.assetId !== AddressZero)
-    .map((p) => p.assetId);
-
-  assets.forEach((asset) => {
-    rebalancedTokens[chainId][asset] = {
-      contract: new Contract(asset, ERC20Abi, provider),
-      decimals: undefined,
-    };
-  });
-});
-
-new Gauge({
-  name: "router_onchain_balance",
-  help: "router_onchain_balance_help",
-  labelNames: ["chainId", "assetId", "signerAddress"] as const,
-  registers: [register],
-  async collect() {
-    await Promise.all(
-      Object.entries(hydrated).map(async ([chainId, provider]) => {
-        // base asset
-        const balance = await provider.getBalance(signer.address);
-        this.set({ chainId, assetId: AddressZero, signerAddress: signer.address }, parseFloat(formatEther(balance)));
-
-        // tokens
-        await Promise.all(
-          Object.entries(rebalancedTokens[chainId] ?? {}).map(async ([assetId, config]) => {
-            const decimals = config.decimals ?? (await config.contract.functions.decimals());
-            rebalancedTokens[chainId][assetId].decimals = decimals;
-            const balance = await config.contract.balanceOf(signer.address);
-            this.set({ chainId, assetId, signerAddress: signer.address }, parseFloat(formatUnits(balance, decimals)));
-          }),
-        );
-      }),
-    );
-  },
+const hydratedProviders = hydrateProviders(config.chainProviders);
+const chainService = new VectorChainReader(hydratedProviders, logger.child({ module: "RouterChainReader" }));
+const messagingService = new NatsRouterMessagingService({
+  signer,
+  logger: logger.child({ module: "NatsRouterMessagingService" }),
+  messagingUrl: config.messagingUrl,
 });
 
 server.addHook("onReady", async () => {
-  const messagingService = new NatsRouterMessagingService({
-    signer,
-    logger: logger.child({ module: "NatsRouterMessagingService" }),
-    messagingUrl: config.messagingUrl,
-  });
   const nodeService = await RestServerNodeService.connect(
     config.nodeUrl,
     logger.child({ module: "RouterNodeService" }),
     evts,
     0,
     true,
-  );
-  const chainService = new VectorChainReader(
-    hydrateProviders(config.chainProviders),
-    logger.child({ module: "RouterChainReader" }),
   );
 
   router = await Router.connect(
@@ -178,8 +136,11 @@ server.addHook("onReady", async () => {
     store,
     messagingService,
     logger,
-    register,
   );
+
+  if (config.autoRebalanceInterval) {
+    startAutoRebalanceTask(config.autoRebalanceInterval, logger, wallet, chainService, hydratedProviders);
+  }
 });
 
 server.get("/ping", async () => {
@@ -193,6 +154,15 @@ server.get("/metrics", async (request, response) => {
   const res = await register.metrics();
   return response.status(200).send(res);
 });
+
+// ADMIN FUNCTIONS
+
+server.post("/auto-rebalance", async (request, response) => {
+  await autoRebalanceTask(logger, wallet, chainService, hydratedProviders);
+  return response.status(200).send({ message: "success" });
+});
+
+// EVENT HANDLERS
 
 server.post(isAlivePath, async (request, response) => {
   evts[EngineEvents.IS_ALIVE].evt!.post(request.body as IsAlivePayload);
@@ -241,6 +211,27 @@ server.post(depositReconciledPath, async (request, response) => {
 
 server.post(requestCollateralPath, async (request, response) => {
   evts[EngineEvents.REQUEST_COLLATERAL].evt!.post(request.body as RequestCollateralPayload);
+  return response.status(200).send({ message: "success" });
+});
+
+server.post(transactionSubmittedPath, async (request, response) => {
+  evts[EngineEvents.TRANSACTION_SUBMITTED].evt!.post(
+    request.body as TransactionSubmittedPayload & { publicIdentifier: string },
+  );
+  return response.status(200).send({ message: "success" });
+});
+
+server.post(transactionMinedPath, async (request, response) => {
+  evts[EngineEvents.TRANSACTION_MINED].evt!.post(
+    request.body as TransactionMinedPayload & { publicIdentifier: string },
+  );
+  return response.status(200).send({ message: "success" });
+});
+
+server.post(transactionFailedPath, async (request, response) => {
+  evts[EngineEvents.TRANSACTION_FAILED].evt!.post(
+    request.body as TransactionFailedPayload & { publicIdentifier: string },
+  );
   return response.status(200).send({ message: "success" });
 });