+++ modules/test-ui/src/App.tsx
@@ -53,6 +53,8 @@ function App() {
         iframeSrc,
         routerPublicIdentifier,
         chainProviders,
+        chainAddresses: config.chainAddresses,
+        messagingUrl: config.messagingUrl,
       });
       await client.init();
       const channelsRes = await client.getStateChannels();
@@ -244,7 +246,6 @@ function App() {
       console.error("Error withdrawing", requestRes.getError());
       return requestRes.getError().message;
     }
-    console.log("requestRes.getValue(): ", requestRes.getValue());
     return requestRes.getValue().signedMessage;
   };
 
@@ -695,7 +696,6 @@ function App() {
                 name="signMessage"
                 onFinish={async (values) => {
                   const signedMessage = await signMessage(values.message);
-                  console.log("signedMessage: ", signedMessage);
                   signMessageForm.setFieldsValue({ signedMessage });
                 }}
                 onFinishFailed={onFinishFailed}