+++ modules/types/src/vectorProvider.ts
@@ -1,4 +1,4 @@
-import { FullTransferState, FullChannelState } from "./channel";
+import { FullTransferState, FullChannelState, ChainAddresses } from "./channel";
 import { ChainProviders } from "./network";
 import { EngineParams, NodeResponses } from "./schemas";
 import { RegisteredTransfer } from "./transferDefinitions";
@@ -16,6 +16,7 @@ export const ChannelRpcMethods = {
   chan_getActiveTransfers: "chan_getActiveTransfers",
   chan_getRegisteredTransfers: "chan_getRegisteredTransfers",
   chan_getTransferState: "chan_getTransferState",
+  chan_getTransfers: "chan_getTransfers",
   chan_setup: "chan_setup",
   chan_sendIsAlive: "chan_sendIsAlive",
   chan_requestSetup: "chan_requestSetup",
@@ -49,6 +50,7 @@ export type ChannelRpcMethodsPayloadMap = {
   [ChannelRpcMethods.chan_getTransferStatesByRoutingId]: EngineParams.GetTransferStatesByRoutingId;
   [ChannelRpcMethods.chan_getActiveTransfers]: EngineParams.GetActiveTransfers;
   [ChannelRpcMethods.chan_getTransferState]: EngineParams.GetTransferState;
+  [ChannelRpcMethods.chan_getTransfers]: EngineParams.GetTransfers;
   [ChannelRpcMethods.chan_getRegisteredTransfers]: EngineParams.GetRegisteredTransfers;
   [ChannelRpcMethods.chan_getChannelStates]: {};
   [ChannelRpcMethods.chan_setup]: EngineParams.Setup;
@@ -61,7 +63,12 @@ export type ChannelRpcMethodsPayloadMap = {
   [ChannelRpcMethods.chan_withdraw]: EngineParams.Withdraw;
   [ChannelRpcMethods.chan_subscribe]: { event: string; once: boolean };
   [ChannelRpcMethods.chan_unsubscribeAll]: {};
-  [ChannelRpcMethods.connext_authenticate]: { signature?: string; chainProviders: ChainProviders };
+  [ChannelRpcMethods.connext_authenticate]: {
+    signature?: string;
+    chainProviders: ChainProviders;
+    chainAddresses?: ChainAddresses;
+    messagingUrl?: string;
+  };
   [ChannelRpcMethods.chan_dispute]: EngineParams.DisputeChannel;
   [ChannelRpcMethods.chan_defund]: EngineParams.DefundChannel;
   [ChannelRpcMethods.chan_disputeTransfer]: EngineParams.DisputeTransfer;
@@ -85,6 +92,7 @@ export type ChannelRpcMethodsResponsesMap = {
   [ChannelRpcMethods.chan_getTransferStateByRoutingId]: FullTransferState | undefined;
   [ChannelRpcMethods.chan_getTransferStatesByRoutingId]: FullTransferState[];
   [ChannelRpcMethods.chan_getActiveTransfers]: FullTransferState[];
+  [ChannelRpcMethods.chan_getTransfers]: FullTransferState[];
   [ChannelRpcMethods.chan_getTransferState]: FullTransferState | undefined;
   [ChannelRpcMethods.chan_getRegisteredTransfers]: RegisteredTransfer[];
   [ChannelRpcMethods.chan_setup]: FullChannelState;