+++ modules/utils/src/test/services/store.ts
@@ -13,6 +13,9 @@ import {
 import { TransactionReceipt, TransactionResponse } from "@ethersproject/abstract-provider";
 
 export class MemoryStoreService implements IEngineStore {
+  getWithdrawalCommitmentByTransactionHash(transactionHash: string): Promise<WithdrawCommitmentJson> {
+    throw new Error("Method not implemented.");
+  }
   saveChannelDispute(
     channel: FullChannelState,
     channelDispute: ChannelDispute,
@@ -40,15 +43,12 @@ export class MemoryStoreService implements IEngineStore {
   //   return Promise.resolve(this.transferDisputes.get(transferAddress));
   // }
   getTransactionByHash(transactionHash: string): Promise<StoredTransaction | undefined> {
-    // console.log("getTransactionByHash not implemented.");
     return Promise.resolve(undefined);
   }
   saveTransactionFailure(channelAddress: string, transactionHash: string, error: string): Promise<void> {
-    // console.log("saveTransactionFailure not implemented.");
     return Promise.resolve(undefined);
   }
   saveTransactionReceipt(channelAddress: string, transaction: TransactionReceipt): Promise<void> {
-    // console.log("saveTransactionReceipt not implemented.");
     return Promise.resolve(undefined);
   }
   saveTransactionResponse(
@@ -56,7 +56,6 @@ export class MemoryStoreService implements IEngineStore {
     transactionReason: TransactionReason,
     response: TransactionResponse,
   ): Promise<void> {
-    // console.log("saveTransactionResponse not implemented.");
     return Promise.resolve(undefined);
   }
   // Map<channelAddress, transferId[]>